CRUD finalizado

+++ src/main/java/com/api/gastapi/controller/PedidoController.java
@@ -8,9 +8,12 @@ import com.api.gastapi.repository.PedidoRepository;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @RestController
@@ -30,6 +33,17 @@ public class PedidoController {
         return repository.findAll().stream().map(PedidoListagemDto::new).toList();
     }
 
+    @GetMapping("/{id_pedido}")
+    public ResponseEntity<Object> buscarPedido(@PathVariable(value = "id_pedido") UUID id_pedido){
+        Optional<PedidoModel> pedidoBuscado = repository.findById(id_pedido);
+
+        if (pedidoBuscado.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuario não encontrado");
+        }
+
+        return ResponseEntity.status(HttpStatus.OK).body(pedidoBuscado.get());
+    }
+
     @PutMapping
     @Transactional
     public void atualizarPedido(@RequestBody @Valid PedidoAtualizacaoDto dados){
@@ -39,7 +53,13 @@ public class PedidoController {
 
     @DeleteMapping("/{id_pedido}")
     @Transactional
-    public void excluirPedido(@PathVariable UUID id_pedido){
+    public ResponseEntity<Object> deletarPedido(@PathVariable(value = "id_pedido") UUID id_pedido) {
+        Optional<PedidoModel> pedidoBuscado = repository.findById(id_pedido);
+
+        if (pedidoBuscado.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuario não encontrado");
+        }
         repository.deleteById(id_pedido);
+        return ResponseEntity.status(HttpStatus.OK).body("Usuario deletado com sucesso!");
     }
 }
